refactoring - seems to be bug-free
