+++ src/components/myApplyances/MyApplyanceContainer.js
@@ -1,8 +1,9 @@
 // parent container of MyApplyanceList (list of my applyances) and MyApplyanceSearch (searching applyances)
 import { useState } from "react"
 import { MyApplyanceList } from "./MyApplyanceList"
-import { MyApplyanceSearch } from "./MyApplyanceSearch"
+// import { MyApplyanceSearch } from "./MyApplyanceSearch"
 import "../applyances/Applyances.css"
+import { ApplyanceSearch } from "../applyances/ApplyanceSearch"
 
 
 export const MyApplyanceContainer = () => {
@@ -14,7 +15,7 @@ export const MyApplyanceContainer = () => {
 
     
         {/* MyApplyanceSearch component allows a user to input search terms by typing into a text field, which calls the mySetSearchTerms function that was passed down and updates the mySearchTerms state variable */}
-        <MyApplyanceSearch mySetterFunction={mySetSearchTerms}/>
+        <ApplyanceSearch mySetterFunction={mySetSearchTerms}/>
 
         {/* MyApplyanceList component receives the mySearchTermState prop from the parent and uses it to filter the list of applyances that it displays. If empty, it displays all my applyances. Otherwise, it displays a filtered list based on the whether the make/model includes the mySearchTerms string */}
         <MyApplyanceList mySearchTermState={mySearchTerms}/>
